clean up measurement component
