+++ package.json
@@ -51,5 +51,5 @@
   "devDependencies": {
     "gh-pages": "^3.1.0"
   },
-  "proxy": "http://20.0.101.107:8080/fhir"
+  "proxy": "http://dev-ehr.ihealth-eng.com/fhir"
 }
