+++ src/App.scss
@@ -42,6 +42,7 @@
   border: 1px solid #e9e9e9;
   font-weight: 400;
   font-size: 12px;
+  text-align: center;
 }
 
 .statisticsTableRow{
@@ -66,6 +67,7 @@
   border: 1px solid #e9e9e9;
   font-weight: 400;
   font-size: 12px;
+  text-align: center;
 }
 
 .bedtime{
@@ -100,4 +102,14 @@
   border-radius: 5px;
   // width: 10em;
   box-shadow: 10px 10px 0 rgba(#523E6B, 0.2);
+}
+
+.bgTable {
+  margin-bottom: 30px;
+  width: '49%';
+  text-align: "center";
+}
+
+.ant-table {
+  text-align: center;
 }
